+++ CHANGELOG.md
@@ -11,6 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [6.3.0] - 2026-08-05
+
+### Added
+
+- Add application-authenticated playlist playlog delivery to Flare.
+- Add application-authenticated campaign playlog delivery to Ignis.
+
 ## [6.1.0] - 2026-08-03
 
 ### Added